requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.2.0b2
azure-ai-projects==1.1.0b2
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.24.0
azure-search-documents==11.6.0b12
azure-storage-blob==12.26.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==45.0.6
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
fastapi==0.116.1
frozenlist==1.7.0
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib-metadata==8.7.0
isodate==0.7.2
jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
lxml==6.0.0
markupsafe==3.0.2
more-itertools==10.7.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.99.9
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
parse==1.20.2
pathable==0.4.4
prance==25.4.8.0
propcache==0.3.2
protobuf==6.31.1
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyee==13.0.0
pyjwt==2.10.1
pylibsrtp==0.12.0
pymeta3==0.5.1
pyopenssl==25.1.0
python-dotenv==1.1.1
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rpds-py==0.27.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
scipy==1.16.1
semantic-kernel==1.35.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
